Create Calendar Project
This Calendar application can perform mainly three operations those are finding the day corresponding to a given date, month and the year which were asked.

After executing the code you can navigate using the arrow keys of the keyboard. For next you should press the key ‘n’ and for previous just press the ‘p’ key.

One of the best features of this application is that it utilizes the method of file handling. with the help of this wonderful feature, you can add some important notes to the specific dates.

The calendar application is mainly based on the console which does not contain graphics. The color used for the days of month is white and the Sundays are highlighted using red color.